Fix error , if entry doesn't exist in primary site
If an entry does not exist in the primary site (e.g. only in a second site), the request for the element requests with the wrong siteId (the siteId from the primary site). The request should also have the current siteId.

(cherry picked from commit 4071479b5cf2bc39094c5771c20256688a83ecaa)